fix: Limit checksum update to each file
Otherwise, sed will just update the first matching checksum. This shouldn't normally be an issue, but the easiest way to add new files is to copy the line of an existing one and update the filename, but if this is done the file will have duplicate checksums and won't work as expected.